+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All changes to this project will be documented in this file.
 
+## [1.2.1] - 2023-08-25
+- Fix progressive upload with upload token and video id
+- Use pascal case for enums
+
 ## [1.2.0] - 2023-06-28
 - Introducing new live streams restream feature
 - Introducing new analytics endpoints